feat: Refactor AttendanceDTO and Attendance classes and jwt cookie auth now works
Refactor the AttendanceDTO and Attendance classes to remove unnecessary code duplication and improve code organization. Remove the default constructor in AttendanceDTO and move the getters and setters to the bottom of the class. Update the Attendance class to use constructor injection for dependencies and improve the naming of the setter method for the attendedAt field. This refactoring improves the readability and maintainability of the codebase.